Rename `format` folder to `model`
