+++ README.md
@@ -70,23 +70,23 @@ There are two core types of nodes: _Text_ and _Element_.
 
 There is a number of block and inline elements our editor can generate.
 
-- [AttachmentNode](./src/format/elements/AttachmentNode.ts)
-- [BookmarkNode](./src/format/elements/BookmarkNode.ts)
-- [ContactNode](./src/format/elements/ContactNode.ts)
-- [CoverageNode](./src/format/elements/CoverageNode.ts)
-- [DividerNode](./src/format/elements/DividerNode.ts)
-- [EmbedNode](./src/format/elements/EmbedNode.ts)
-- [GalleryNode](./src/format/elements/GalleryNode.ts)
-- [HeadingNode](./src/format/elements/HeadingNode.ts)
-- [HtmlNode](./src/format/elements/HtmlNode.ts)
-- [ImageNode](./src/format/elements/ImageNode.ts)
-- [LinkNode](./src/format/elements/LinkNode.ts)
-- [ListNode](./src/format/elements/ListNode.ts)
-- [MentionNode](./src/format/elements/MentionNode.ts)
-- [ParagraphNode](./src/format/elements/ParagraphNode.ts)
-- [PlaceholderNode](./src/format/elements/PlaceholderNode.ts)
-- [QuoteNode](./src/format/elements/QuoteNode.ts)
-- [VideoNode](./src/format/elements/VideoNode.ts)
+- [AttachmentNode](src/model/elements/AttachmentNode.ts)
+- [BookmarkNode](src/model/elements/BookmarkNode.ts)
+- [ContactNode](src/model/elements/ContactNode.ts)
+- [CoverageNode](src/model/elements/CoverageNode.ts)
+- [DividerNode](src/model/elements/DividerNode.ts)
+- [EmbedNode](src/model/elements/EmbedNode.ts)
+- [GalleryNode](src/model/elements/GalleryNode.ts)
+- [HeadingNode](src/model/elements/HeadingNode.ts)
+- [HtmlNode](src/model/elements/HtmlNode.ts)
+- [ImageNode](src/model/elements/ImageNode.ts)
+- [LinkNode](src/model/elements/LinkNode.ts)
+- [ListNode](src/model/elements/ListNode.ts)
+- [MentionNode](src/model/elements/MentionNode.ts)
+- [ParagraphNode](src/model/elements/ParagraphNode.ts)
+- [PlaceholderNode](src/model/elements/PlaceholderNode.ts)
+- [QuoteNode](src/model/elements/QuoteNode.ts)
+- [VideoNode](src/model/elements/VideoNode.ts)
 
 ## Example document
