+++ src/model/Node.ts
