+++ src/model/Text.ts
