+++ src/model/common/OEmbedInfo.ts
