+++ src/model/elements/AttachmentNode.ts
