+++ src/model/elements/BookmarkNode.ts
