+++ src/model/elements/CoverageNode.ts
