+++ src/model/elements/GalleryNode.ts
