+++ src/model/elements/ImageNode.ts
