+++ src/model/elements/LinkNode.ts
