+++ src/model/elements/MentionNode.ts
