+++ src/model/elements/QuoteNode.ts
