+++ src/model/elements/StoryBookmarkNode.ts
