+++ src/model/validation/isArrayOf.ts
