+++ src/model/validation/isBoolean.ts
