+++ src/model/validation/isEnum.ts
