+++ src/model/validation/isNonEmptyString.ts
