+++ src/model/validation/isNonZeroInteger.ts
