+++ src/model/validation/isObject.ts
