+++ src/model/validation/isString.ts
