+++ src/model/validation/isUuid.ts
